refactor
- 회원가입 및 로그인 시 JSON 형식
- 프로젝트 변수 설정 파일(application.yam) 분리
- 변수 이름 변경

+++ voice-me-sing/src/main/java/com/wandookong/voice_me_sing/repository/UserRepository.java
@@ -3,7 +3,8 @@ package com.wandookong.voice_me_sing.repository;
 import com.wandookong.voice_me_sing.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
     boolean existsByUsername(String username);
 
     UserEntity findByUsername(String username);
